refactor: nickname 필드 네이밍 수정
nickname -> nickName
